requirements.txt
streamlit>=1.28.0
requests>=2.31.0
# requests is used for both GitHub Codespaces API and GitHub Storage API
python-dotenv>=1.0.0
python-dateutil>=2.8.2
# python-dateutil provides advanced datetime parsing and manipulation capabilities

